+++ CHANGELOG.md
@@ -49,6 +49,7 @@ v0.20 (trunk)
 * Add support of XEP-0393: Message Styling
 * Add a BundleCapabilityResolver to link OMEMO bundles with the capability/info table
 * Add pagination for Pictures and Links in the Contact and Room drawers
+* Implement affiliations modification and ban/unban in Rooms
 
 v0.19
 ---------------------------